+++ Adapters/Bidease/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Bidease iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationBidease', '2.2.1.0'
+pod 'CASMediationBidease', '2.2.5.0'
 ```
+
+### 2.2.5.0
+- Certified with Bidease - 2.2.5
 
 ### 2.2.1.0
 - Certified with Bidease - 2.2.1