+++ Adapters/DisplayIO/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Display IO iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationDisplayIO', '4.7.3.0'
+pod 'CASMediationDisplayIO', '4.7.4.0'
 ```
+
+### 4.7.4.0
+- Certified with DisplayIO - 4.7.4
 
 ### 4.7.3.0
 - Certified with DisplayIO - 4.7.3