+++ Adapters/Moloco/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Moloco iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationMoloco', '4.7.0.0'
+pod 'CASMediationMoloco', '4.8.0.0'
 ```
+
+### 4.8.0.0
+- Certified with Moloco - 4.8.0
 
 ### 4.7.0.0
 - Certified with Moloco - 4.7.0